+++ windows/flutter/generated_plugins.cmake
@@ -8,6 +8,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   file_selector_windows
   flutter_onnxruntime
   flutter_secure_storage_windows
+  gal
   hotkey_manager_windows
   irondash_engine_context
   media_kit_libs_windows_audio